cli: Phase D — resize and expand verbs
resize IMG@N --size BYTES is a thin CLI over fs::resize_filesystem_for,
which dispatches in-place to whichever per-FS resize matches the on-
disk magic (FAT/NTFS/exFAT/HFS/HFS+/ext/btrfs/SFS/PFS3/AFFS/EFS).

expand IMG@N --size --block-size --output handles the classic-HFS-only
expand-block-size flow: clone the source HFS volume into a freshly
formatted blank of the chosen size + allocation block size, then wrap
the result in a new APM disk image.

To make expand shareable, gui/expand_hfs_dialog.rs::run_expand /
summarize_source / ExpandSource / BLOCK_SIZE_CHOICES move into
src/model/hfs_expand_runner.rs. The GUI dialog now calls
start_hfs_expand and re-exports summarize_source so existing call
sites keep working.
