+++ docs/cli-todo.md
@@ -17,7 +17,7 @@ it in the binary yet?"
 | **A** Rename + flat verbs + infrastructure | **shipped** | `5d0c431..7bb1385` — binary rename, module split, IMG@N parser, `--format` envelope, logging flags, exit codes, flat verbs (`new`, `ls`, `put`, `get`, `rm`, `mkdir`, `fsck`, `shrink`, `inspect`, `show {partmap,fs-info,chd-info,devices}`, `completions`, `install-completions`), manpage example, `--help` snapshot tests |
 | **B** Generalize FS verbs + globbing | **shipped** | `b627eca..f40ed4c` — ls/put/get/rm/mkdir/fsck route through `open_filesystem` / `open_editable_filesystem`, bash-equivalent globs (`*`, `?`, `[...]`, `**`, `{a,b}`) with `--include`/`--exclude` (exclude-wins). Blank creators for FAT/EFS/AFFS still pending; `new --fs fat/efs/affs` gated on those landing. |
 | **C** Big-ticket verbs | **4 of 5 shipped** | `9c465d0` — `backup`, `restore`, `write` (background progress pump, `--yes` device safety gate, fs-aware `inspect` text/json/yaml). `convert` (this slice) lifts `run_bulk_convert` into `src/model/bulk_convert_runner.rs` and adds `rb-cli convert IN OUT --format FMT` over it; GUI Bulk Convert dialog now calls into the shared runner. |
-| **D** Batch + resize/expand cluster | **2 of 3 shipped** | `6a59121` — `rb-cli batch SCRIPT.json` with mkdir/put/rm ops, `--dry-run`, `--continue-on-error`. `batch-template` (this slice) — add-to-existing-image mode (host dir → mkdir/put script, builtin extension→type/creator table, include/exclude globs). **Deferred**: `resize`, `expand` (non-SGI filesystems), whole-volume `format` op + multi-partition `disk` block (gated on FAT/EFS/AFFS blank creators from Phase B). |
+| **D** Batch + resize/expand cluster | **3 of 3 shipped (+1 deferred sub-item)** | `6a59121` — `rb-cli batch SCRIPT.json`, `batch-template`. `resize IMG@N --size BYTES` routes through `fs::resize_filesystem_for` (FAT/NTFS/exFAT/HFS/HFS+/ext/btrfs/SFS/PFS3/AFFS/EFS). `expand IMG@N --size --block-size --output` lifts the GUI's HFS expand-block-size dialog into `src/model/hfs_expand_runner.rs` and exposes it as a CLI. **Deferred sub-item only**: whole-volume `format` op + multi-partition `disk` block in batch (gated on FAT/EFS/AFFS blank creators from Phase B). |
 | **E** Optical | **pending** | `optical rip / convert / browse / extract` — needs `opticaldiscs`/`cd-da-reader` plumbing wired to a new verb tree. |
 | **F** Interactive `terminal` REPL | **pending** | `rustyline`/`reedline` REPL on top of the one-shot verbs. |
 | **G** Config file | **shipped** | `15ce270` — INI loader, XDG-aware `default_path()`, `rb-cli config {init,path,show}`. Per-verb flag resolution (CLI > config > default) wired for: `[defaults] log-level / progress / color` (globals applied in `logging::install`); `[backup] format / checksum`. New `--config PATH` global flag overrides the default location. Remaining per-verb keys (e.g. `[fsck] prompt-timeout`) will be threaded as the underlying flag becomes load-bearing. |
