+++ src/cli/verbs/mod.rs
@@ -12,6 +12,7 @@ pub mod batch_template;
 pub mod completions;
 pub mod config;
 pub mod convert;
+pub mod expand;
 pub mod fsck;
 pub mod get;
 pub mod inspect;
@@ -19,6 +20,7 @@ pub mod ls;
 pub mod mkdir;
 pub mod new;
 pub mod put;
+pub mod resize;
 pub mod restore;
 pub mod rm;
 pub mod show;
