+++ src/gui/expand_hfs_dialog.rs
@@ -8,65 +8,18 @@
 //!
 //! Source disk is opened read-only; the output is always a brand-new file.
 
-use std::io::{Cursor, Seek, SeekFrom};
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
-use rusty_backup::fs::hfs::{create_blank_hfs_sized, HfsFilesystem};
-use rusty_backup::fs::hfs_clone::{
-    clone_hfs_volume, emit_apm_disk_with_hfs, CloneReport, EmitReport,
+use rusty_backup::model::hfs_expand_runner::{
+    max_volume_for_block_size, start_hfs_expand, suggest_block_size, ExpandSource,
+    BLOCK_SIZE_CHOICES,
 };
-use rusty_backup::model::source_reader::open_read;
 use rusty_backup::model::status::ExpandStatus;
 
 use super::progress::LogPanel;
 
-/// Allocation block sizes the user can pick. Each maps to a `2 GB`-class
-/// ceiling defined by the 65535-block u16 in HFS.
-const BLOCK_SIZE_CHOICES: &[u32] = &[4096, 8192, 16384, 32768, 65536];
-
-/// Maximum HFS volume size for a given allocation block size.
-fn max_volume_for_block_size(bs: u32) -> u64 {
-    65535u64 * bs as u64
-}
-
-/// Suggest the smallest block size whose ceiling can hold `target_bytes`.
-/// Falls back to the largest available size if nothing fits comfortably.
-fn suggest_block_size(target_bytes: u64) -> u32 {
-    for &bs in BLOCK_SIZE_CHOICES {
-        if max_volume_for_block_size(bs) >= target_bytes {
-            return bs;
-        }
-    }
-    *BLOCK_SIZE_CHOICES.last().unwrap()
-}
-
-/// Source-volume summary the dialog displays + uses to validate inputs.
-#[derive(Debug, Clone)]
-pub struct ExpandSource {
-    /// Path to the source disk file (image or device path).
-    pub source_path: PathBuf,
-    /// Byte offset of the source HFS partition on the disk.
-    pub partition_offset: u64,
-    /// Source partition size in bytes (block_count × 512).
-    pub partition_size: u64,
-    /// Source allocation block size (drAlBlkSiz), informational.
-    pub source_block_size: u32,
-    /// Source catalog file size (bytes). Used to size the target catalog so
-    /// it can hold every record from the source without running out of nodes.
-    pub source_catalog_size: u32,
-    /// Source extents-overflow file size (bytes). Used the same way for the
-    /// target's extents-overflow B-tree.
-    pub source_extents_size: u32,
-    /// Source allocated bytes (used = (total_blocks - free_blocks) × block_size).
-    pub used_bytes: u64,
-    /// Source file count, informational.
-    pub file_count: u64,
-    /// Source directory count, informational.
-    pub dir_count: u64,
-    /// Source volume name (carried into the new volume).
-    pub volume_name: String,
-}
+pub use rusty_backup::model::hfs_expand_runner::summarize_source;
 
 pub struct ExpandHfsDialog {
     pub open: bool,
@@ -295,20 +248,9 @@ impl ExpandHfsDialog {
     }
 
     fn spawn_worker(&mut self, output: PathBuf, log: &mut LogPanel) {
-        let status = Arc::new(Mutex::new(ExpandStatus {
-            finished: false,
-            error: None,
-            log_messages: Vec::new(),
-            current_step: "Starting…".to_string(),
-            clone_report: None,
-            emit_report: None,
-        }));
-        self.status = Some(Arc::clone(&status));
-
         let source = self.source.clone();
         let target_size = self.target_size_mib as u64 * 1024 * 1024;
         let target_bs = self.target_block_size;
-
         log.info(format!(
             "Expanding HFS volume '{}' to {} MiB at {} KiB blocks -> {}",
             source.volume_name,
@@ -316,24 +258,7 @@ impl ExpandHfsDialog {
             target_bs / 1024,
             output.display()
         ));
-
-        std::thread::spawn(move || {
-            let _wake = rusty_backup::os::wakelock::acquire("Rusty Backup: expand HFS volume");
-            let result = run_expand(source, target_size, target_bs, output, &status);
-            let mut s = status.lock().expect("expand status lock");
-            s.finished = true;
-            match result {
-                Ok((clone, emit)) => {
-                    s.clone_report = Some(clone);
-                    s.emit_report = Some(emit);
-                    s.current_step = "Done.".into();
-                }
-                Err(e) => {
-                    s.error = Some(e.to_string());
-                    s.current_step = "Failed.".into();
-                }
-            }
-        });
+        self.status = Some(start_hfs_expand(source, target_size, target_bs, output));
     }
 }
 
@@ -356,190 +281,4 @@ fn sanitize_filename(s: &str) -> String {
     } else {
         trimmed
     }
-}
-
-/// Worker body for one expand run. Returns the clone+emit reports on success.
-fn run_expand(
-    source: ExpandSource,
-    target_size_bytes: u64,
-    target_block_size: u32,
-    output_path: PathBuf,
-    status: &Arc<Mutex<ExpandStatus>>,
-) -> anyhow::Result<(CloneReport, EmitReport)> {
-    let push = |s: &str| {
-        if let Ok(mut g) = status.lock() {
-            g.log_messages.push(s.to_string());
-        }
-    };
-    let step = |s: &str| {
-        if let Ok(mut g) = status.lock() {
-            g.current_step = s.to_string();
-        }
-    };
-
-    step("Building blank target volume…");
-    // Size the target B-trees from the source's actual on-disk file sizes,
-    // with 50% headroom on the catalog (split-based insertion can leave
-    // some nodes underfull). The blank-volume builder appends map nodes
-    // (BTNodeKind=2) to extend the node-allocation bitmap past the
-    // header-node bitmap's 2048-bit cap, so requesting a >1 MiB catalog
-    // is fine. We still cap at HFS_MAX_BTREE_FILE_SIZE (16 MiB) because
-    // larger trees are not needed for any realistic HFS volume.
-    let catalog_min = source
-        .source_catalog_size
-        .saturating_mul(3)
-        .saturating_div(2)
-        .min(rusty_backup::fs::hfs::HFS_MAX_BTREE_FILE_SIZE);
-    let extents_min = source
-        .source_extents_size
-        .min(rusty_backup::fs::hfs::HFS_MAX_BTREE_FILE_SIZE);
-    let mut target_buf = create_blank_hfs_sized(
-        target_size_bytes,
-        target_block_size,
-        &source.volume_name,
-        extents_min,
-        catalog_min,
-    )
-    .map_err(|e| anyhow::anyhow!("create_blank_hfs failed: {e}"))?;
-    push(&format!(
-        "Target B-trees sized: catalog ≥ {} KiB (source {} KiB), extents-overflow ≥ {} KiB (source {} KiB)",
-        catalog_min / 1024,
-        source.source_catalog_size / 1024,
-        extents_min / 1024,
-        source.source_extents_size / 1024
-    ));
-    push(&format!(
-        "Allocated {} MiB target image at {} KiB blocks",
-        target_buf.len() / (1024 * 1024),
-        target_block_size / 1024
-    ));
-
-    step("Reading source catalog…");
-    let mut source_buffered =
-        open_read(&source.source_path).map_err(|e| anyhow::anyhow!("open source: {e}"))?;
-    let source_disk_size = source_buffered
-        .seek(SeekFrom::End(0))
-        .map_err(|e| anyhow::anyhow!("source size: {e}"))?;
-    source_buffered
-        .seek(SeekFrom::Start(0))
-        .map_err(|e| anyhow::anyhow!("source rewind: {e}"))?;
-
-    let mut source_fs = HfsFilesystem::open(source_buffered, source.partition_offset)
-        .map_err(|e| anyhow::anyhow!("open source HFS: {e}"))?;
-
-    // Capture which fsck issue codes the source already exhibits — pre-
-    // existing quirks (e.g. system files with embedded null bytes in their
-    // catalog names) faithfully copied to the target shouldn't fail the
-    // post-clone verification.
-    let source_issue_codes: std::collections::HashSet<String> = source_fs
-        .fsck()
-        .map(|r| r.errors.iter().map(|e| e.code.clone()).collect())
-        .unwrap_or_default();
-    if !source_issue_codes.is_empty() {
-        push(&format!(
-            "Source has pre-existing fsck issue codes: {:?} (will be ignored on target)",
-            source_issue_codes
-        ));
-    }
-
-    step("Cloning files…");
-    let target_cursor = Cursor::new(&mut target_buf);
-    let mut target_fs = HfsFilesystem::open(target_cursor, 0)
-        .map_err(|e| anyhow::anyhow!("open target HFS: {e}"))?;
-    let clone_report = clone_hfs_volume(&mut source_fs, &mut target_fs)
-        .map_err(|e| anyhow::anyhow!("clone failed: {e}"))?;
-    push(&format!(
-        "Copied {} file(s) and {} directory(ies); {} bytes data + {} bytes resource fork",
-        clone_report.files_copied,
-        clone_report.dirs_copied,
-        clone_report.data_bytes_copied,
-        clone_report.rsrc_bytes_copied
-    ));
-
-    step("Verifying target with fsck…");
-    {
-        // Re-open target read-only to fsck the cloned-and-synced bytes.
-        let mut verify_fs = HfsFilesystem::open(Cursor::new(&target_buf), 0)
-            .map_err(|e| anyhow::anyhow!("reopen target for fsck: {e}"))?;
-        let result = verify_fs
-            .fsck()
-            .map_err(|e| anyhow::anyhow!("fsck error: {e}"))?;
-        let (new_issues, inherited): (Vec<_>, Vec<_>) = result
-            .errors
-            .iter()
-            .partition(|i| !source_issue_codes.contains(&i.code));
-        for issue in &inherited {
-            push(&format!(
-                "fsck (inherited from source): [{}] {}",
-                issue.code, issue.message
-            ));
-        }
-        for issue in &new_issues {
-            push(&format!("fsck: [{}] {}", issue.code, issue.message));
-        }
-        if !new_issues.is_empty() {
-            let codes: Vec<String> = new_issues.iter().map(|e| e.code.clone()).collect();
-            anyhow::bail!(
-                "target fsck reported {} new error(s) not present in source: {:?}",
-                codes.len(),
-                codes
-            );
-        }
-    }
-
-    step("Writing new APM disk image…");
-    let output_file = std::fs::OpenOptions::new()
-        .read(true)
-        .write(true)
-        .create(true)
-        .truncate(true)
-        .open(&output_path)
-        .map_err(|e| anyhow::anyhow!("create output: {e}"))?;
-    let mut source_for_apm =
-        open_read(&source.source_path).map_err(|e| anyhow::anyhow!("reopen source: {e}"))?;
-    let emit_report = {
-        // emit_apm_disk_with_hfs needs Read+Write+Seek on the writer.
-        let mut writer = output_file;
-        emit_apm_disk_with_hfs(
-            &mut source_for_apm,
-            source_disk_size,
-            &target_buf,
-            &mut writer,
-        )
-        .map_err(|e| anyhow::anyhow!("emit APM: {e}"))?
-    };
-    push(&format!(
-        "Wrote {} MiB output: {} driver(s) preserved, HFS at block {} ({} blocks)",
-        emit_report.total_bytes / (1024 * 1024),
-        emit_report.drivers_copied,
-        emit_report.hfs_start_block,
-        emit_report.hfs_block_count
-    ));
-
-    Ok((clone_report, emit_report))
-}
-
-/// Quickly summarise a source HFS volume given a path + partition byte
-/// offset. Used by the inspect tab to populate the dialog before showing it.
-pub fn summarize_source(
-    source_path: &std::path::Path,
-    partition_offset: u64,
-    partition_size: u64,
-) -> anyhow::Result<ExpandSource> {
-    let reader = open_read(source_path)?;
-    let summary = HfsFilesystem::open(reader, partition_offset)
-        .map_err(|e| anyhow::anyhow!("open HFS: {e}"))?
-        .volume_summary();
-    Ok(ExpandSource {
-        source_path: source_path.to_path_buf(),
-        partition_offset,
-        partition_size,
-        source_block_size: summary.block_size,
-        source_catalog_size: summary.catalog_file_size,
-        source_extents_size: summary.extents_file_size,
-        used_bytes: summary.used_bytes,
-        file_count: summary.file_count as u64,
-        dir_count: summary.folder_count as u64,
-        volume_name: summary.volume_name,
-    })
 }
