+++ src/model/mod.rs
@@ -11,6 +11,7 @@ pub mod chd_expand_runner;
 pub mod edit_queue;
 pub mod export_runner;
 pub mod fsck_runner;
+pub mod hfs_expand_runner;
 pub mod min_size_runner;
 pub mod partition_editor;
 pub mod physical_write_runner;
